New Parametric Model with Sensor Tray
MAX30101 PCB added to model along with a modified lid and a tray to secure the PCB. The tray has four risers with through holes for m3 screws to pass through and two risers with self-threading holes for securing the MAX30101 PCB. First secure the MAX30101 to the tray, and then secure the tray to the lid. Compatible with existing case bodies and parametric case customizer script.